temp: jin-lee

+++ srcs/exec/exec_cmd.c
@@ -6,7 +6,7 @@
 /*   By: jin-lee <jin-lee@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/03 13:12:06 by jin-lee           #+#    #+#             */
-/*   Updated: 2022/02/03 14:47:56 by jin-lee          ###   ########.fr       */
+/*   Updated: 2022/02/04 16:25:56 by jin-lee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ char	**get_envp(t_elist *elist)
 	idx = 0;
 	curr = elist->head;
 	envp = (char **)malloc(sizeof(char *) * (elist->count + 1));
-	while (idx < elist->count)
+	while (idx < (int)elist->count)
 	{
 		temp = ft_strdup(curr->key);
 		temp = ft_strjoin(temp, "=");
